python/bear_detection/requirement.txt
ultralytics
opencv-python
tensorflow-directml
torch-directml
numpy
